+++ unpacked/_next/static/chunks/pages/admin/settings.js
@@ -6,40 +6,41 @@
       n.r(t),
         n.d(t, {
           __N_SSP: function () {
-            return N;
-          },
-          default: function () {
             return E;
           },
+          default: function () {
+            return D;
+          },
         });
-      var r = n(36112),
-        i = n(35017),
+      var i = n(36112),
+        r = n(35017),
         s = n(82256),
         a = n(90166),
         o = n(67558),
         l = n(45754),
-        c = n(73582),
+        c = n(7144),
+        d = n(73582),
         u = n(82473),
-        d = n(25494),
-        g = n(19841),
+        g = n(25494),
+        p = n(19841),
         f = n(29124),
-        p = n.n(f),
-        h = n(70671),
-        x = n(32004),
-        m = n(94968),
-        b = n(33562),
-        j = n(4748),
-        y = n(81942),
-        v = n(35250);
-      function w(e, t) {
+        h = n.n(f),
+        x = n(70671),
+        b = n(32004),
+        j = n(94968),
+        m = n(33562),
+        y = n(4748),
+        v = n(81942),
+        w = n(35250);
+      function S(e, t) {
         var n = Object.keys(e);
         if (Object.getOwnPropertySymbols) {
-          var r = Object.getOwnPropertySymbols(e);
+          var i = Object.getOwnPropertySymbols(e);
           t &&
-            (r = r.filter(function (t) {
+            (i = i.filter(function (t) {
               return Object.getOwnPropertyDescriptor(e, t).enumerable;
             })),
-            n.push.apply(n, r);
+            n.push.apply(n, i);
         }
         return n;
       }
@@ -47,12 +48,12 @@
         for (var t = 1; t < arguments.length; t++) {
           var n = null != arguments[t] ? arguments[t] : {};
           t % 2
-            ? w(Object(n), !0).forEach(function (t) {
-                (0, r.Z)(e, t, n[t]);
+            ? S(Object(n), !0).forEach(function (t) {
+                (0, i.Z)(e, t, n[t]);
               })
             : Object.getOwnPropertyDescriptors
               ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
-              : w(Object(n)).forEach(function (t) {
+              : S(Object(n)).forEach(function (t) {
                   Object.defineProperty(
                     e,
                     t,
@@ -62,125 +63,125 @@
         }
         return e;
       }
-      function S(e) {
+      function k(e) {
         var t,
           n = e.conversationTTL,
-          r = (0, h.Z)();
+          i = (0, x.Z)();
         return (
           (t =
             null === n
-              ? r.formatMessage(T.retentionInfinite)
+              ? i.formatMessage(_.retentionInfinite)
               : 7776e3 === n
-                ? r.formatMessage(T.retentionNinetyDays)
+                ? i.formatMessage(_.retentionNinetyDays)
                 : 31536e3 === n
-                  ? r.formatMessage(T.retentionOneYear)
-                  : r.formatMessage(T.retentionCustomDays, {
+                  ? i.formatMessage(_.retentionOneYear)
+                  : i.formatMessage(_.retentionCustomDays, {
                       num: Math.floor(n / 86400),
                     })),
-          (0, v.jsxs)(y.vi, {
+          (0, w.jsxs)(v.vi, {
             children: [
-              (0, v.jsx)("p", {
+              (0, w.jsx)("p", {
                 className: "inline-flex items-center text-base",
-                children: (0, v.jsx)(x.Z, O({}, T.conversationRetention)),
+                children: (0, w.jsx)(b.Z, O({}, _.conversationRetention)),
               }),
-              (0, v.jsx)(a.u, {
-                label: r.formatMessage(T.retentionSettingTooltip),
+              (0, w.jsx)(a.u, {
+                label: i.formatMessage(_.retentionSettingTooltip),
                 children: t,
               }),
             ],
           })
         );
       }
-      function k(e) {
+      function Z(e) {
         var t = e.title,
           n = e.description,
-          r = e.children,
-          i = e.upsellEnterprise,
-          s = void 0 !== i && i;
-        return (0, v.jsx)(a.E, {
-          label: (0, v.jsx)(x.Z, O({}, T.enterpriseRequiredMessage)),
+          i = e.children,
+          r = e.upsellEnterprise,
+          s = void 0 !== r && r;
+        return (0, w.jsx)(a.E, {
+          label: (0, w.jsx)(b.Z, O({}, _.enterpriseRequiredMessage)),
           side: "top",
           sideOffset: 4,
           disabled: !s,
           className: "max-w-xl",
-          children: (0, v.jsxs)(y.i9, {
-            className: (0, g.default)(s && "text-token-text-tertiary"),
+          children: (0, w.jsxs)(v.i9, {
+            className: (0, p.default)(s && "text-token-text-tertiary"),
             children: [
-              (0, v.jsxs)(y.A4, {
+              (0, w.jsxs)(v.A4, {
                 className: "flex items-center gap-3",
                 children: [
-                  (0, v.jsx)(x.Z, O({}, t)),
+                  (0, w.jsx)(b.Z, O({}, t)),
                   s &&
-                    (0, v.jsx)("span", {
+                    (0, w.jsx)("span", {
                       className:
                         "rounded-xl bg-token-surface-tertiary px-2 pb-0.5 text-sm",
-                      children: (0, v.jsx)(x.Z, O({}, T.enterpriseUpsellPill)),
+                      children: (0, w.jsx)(b.Z, O({}, _.enterpriseUpsellPill)),
                     }),
                 ],
               }),
-              n && (0, v.jsx)(y.Li, { children: (0, v.jsx)(x.Z, O({}, n)) }),
-              r,
+              n && (0, w.jsx)(v.Li, { children: (0, w.jsx)(b.Z, O({}, n)) }),
+              i,
             ],
           }),
         });
       }
-      function Z(e) {
+      function M(e) {
         var t,
           n = e.workspaceId,
-          r = e.label,
-          i = e.setting,
+          i = e.label,
+          r = e.setting,
           s = e.defaultValue,
           a = e.values,
           o = e.disabled,
-          c = (0, h.Z)(),
-          g = (0, u.NL)(),
-          f = (0, d.D)({
+          c = (0, x.Z)(),
+          d = (0, u.NL)(),
+          p = (0, g.D)({
             mutationFn: function (e) {
               var t = e.setting,
-                r = e.value;
-              return l.ZP.setWorkspaceShareSetting(n, t, r);
+                i = e.value;
+              return l.ZP.setWorkspaceShareSetting(n, t, i);
             },
             onSettled: function () {
-              g.invalidateQueries({ queryKey: ["workspace-settings"] });
+              d.invalidateQueries({ queryKey: ["workspace-settings"] });
             },
             onError: function () {
-              j.m.danger(c.formatMessage(T.featureSettingsUpdateFailed));
+              y.m.danger(c.formatMessage(_.featureSettingsUpdateFailed));
             },
           }),
-          p = f.variables,
-          m = f.mutate;
-        return (0, v.jsxs)(y.vi, {
+          f = p.variables,
+          h = p.mutate;
+        return (0, w.jsxs)(v.vi, {
           children: [
-            (0, v.jsx)("p", {
+            (0, w.jsx)("p", {
               className: "text-md inline-flex items-center",
-              children: (0, v.jsx)(x.Z, O({}, r)),
+              children: (0, w.jsx)(b.Z, O({}, i)),
             }),
-            (0, v.jsxs)(b.Z.Root, {
+            (0, w.jsxs)(m.Z.Root, {
               defaultValue:
-                null !== (t = null == p ? void 0 : p.value) && void 0 !== t
+                null !== (t = null == f ? void 0 : f.value) && void 0 !== t
                   ? t
                   : s,
               onValueChange: function (e) {
-                m({ setting: i, value: e });
+                h({ setting: r, value: e });
               },
               disabled: void 0 !== o && o,
               children: [
-                (0, v.jsxs)(b.Z.Trigger, {
+                (0, w.jsxs)(m.Z.Trigger, {
                   className: "text-md focus-visible:border-none",
                   children: [
-                    (0, v.jsx)(b.Z.Value, {}),
-                    (0, v.jsx)(b.Z.Icon, {}),
+                    (0, w.jsx)(m.Z.Value, {}),
+                    (0, w.jsx)(m.Z.Icon, {}),
                   ],
                 }),
-                (0, v.jsx)(b.Z.Portal, {
-                  children: (0, v.jsx)(b.Z.Content, {
+                (0, w.jsx)(m.Z.Portal, {
+                  children: (0, w.jsx)(m.Z.Content, {
                     children: a.map(function (e, t) {
-                      return (0, v.jsx)(
-                        b.Z.Item,
+                      return (0, w.jsx)(
+                        m.Z.Item,
                         {
                           className: "text-md",
                           value: e.value,
-                          children: (0, v.jsx)(x.Z, O({}, e.display)),
+                          children: (0, w.jsx)(b.Z, O({}, e.display)),
                         },
                         t
                       );
@@ -195,64 +196,81 @@
       function P(e) {
         var t = e.workspaceId,
           n = e.data,
-          r = (0, c.ec)(c.F_.isTeamPlan);
-        return (0, v.jsxs)(y.$V, {
+          i = (0, d.ec)(d.F_.isTeamPlan),
+          r = (0, c.t)();
+        return (0, w.jsxs)(v.$V, {
           children: [
-            r && (0, v.jsx)(y.W8, {}),
-            (0, v.jsx)(k, {
-              title: T.sharingSettingsTitle,
-              upsellEnterprise: r,
-              children: (0, v.jsx)(y.qQ, {
-                children: (0, v.jsx)(Z, {
+            i && (0, w.jsx)(v.W8, {}),
+            (0, w.jsx)(Z, {
+              title: _.workspaceDetails,
+              children: (0, w.jsx)(
+                b.Z,
+                O(
+                  O({}, _.organizationIdLabel),
+                  {},
+                  {
+                    values: {
+                      organizationId:
+                        null == r ? void 0 : r.data.organizationId,
+                    },
+                  }
+                )
+              ),
+            }),
+            (0, w.jsx)(Z, {
+              title: _.sharingSettingsTitle,
+              upsellEnterprise: i,
+              children: (0, w.jsx)(v.qQ, {
+                children: (0, w.jsx)(M, {
                   workspaceId: t,
-                  label: T.chatSharingTitle,
+                  label: _.chatSharingTitle,
                   setting: s.bb.CHAT_SHARE_SETTINGS,
-                  disabled: r,
+                  disabled: i,
                   defaultValue: n.share_settings.chat_share_setting,
                   values: [
                     {
-                      display: T.workspaceMembersOnlySelect,
+                      display: _.workspaceMembersOnlySelect,
                       value: s.Gz.WORKSPACE_ONLY,
                     },
-                    { display: T.noOneSelect, value: s.Gz.NO_ONE },
+                    { display: _.noOneSelect, value: s.Gz.NO_ONE },
                   ],
                 }),
               }),
             }),
-            (0, v.jsx)(k, {
-              title: T.retentionPolicyTitle,
-              description: T.retentionChange,
-              upsellEnterprise: r,
-              children: (0, v.jsx)(S, { conversationTTL: n.conversation_ttl }),
+            (0, w.jsx)(Z, {
+              title: _.retentionPolicyTitle,
+              description: _.retentionChange,
+              upsellEnterprise: i,
+              children: (0, w.jsx)(k, { conversationTTL: n.conversation_ttl }),
             }),
           ],
         });
       }
-      function M(e) {
+      function T(e) {
         var t = e.currentWorkspaceId,
           n = (0, o.C)(t),
-          r = n.data,
+          i = n.data,
           s = n.isLoading,
-          a = (0, h.Z)();
-        return (0, v.jsxs)(v.Fragment, {
+          a = (0, x.Z)();
+        return (0, w.jsxs)(w.Fragment, {
           children: [
-            (0, v.jsx)(p(), {
-              children: (0, v.jsx)("title", {
-                children: a.formatMessage(T.title),
+            (0, w.jsx)(h(), {
+              children: (0, w.jsx)("title", {
+                children: a.formatMessage(_.title),
               }),
             }),
-            (0, v.jsx)(i.yG, { title: a.formatMessage(T.title) }),
-            (0, v.jsx)(i.hb, {
+            (0, w.jsx)(r.yG, { title: a.formatMessage(_.title) }),
+            (0, w.jsx)(r.hb, {
               showSpinner: s,
               children:
-                void 0 === r
+                void 0 === i
                   ? null
-                  : (0, v.jsx)(P, { workspaceId: t, data: r }),
+                  : (0, w.jsx)(P, { workspaceId: t, data: i }),
             }),
           ],
         });
       }
-      var T = (0, m.vU)({
+      var _ = (0, j.vU)({
         title: { id: "workspaceSettings.title", defaultMessage: "Settings" },
         retentionPolicyTitle: {
           id: "workspaceSettings.retentionPolicyTitle",
@@ -265,6 +283,14 @@
         chatSharingTitle: {
           id: "workspaceSettings.chatSharingTitle",
           defaultMessage: "Chats can be shared with...",
+        },
+        workspaceDetails: {
+          id: "workspaceSettings.workspaceDetails",
+          defaultMessage: "Workspace details",
+        },
+        organizationIdLabel: {
+          id: "workspaceSettings.organizationIdLabel",
+          defaultMessage: "Organization ID: {organizationId}",
         },
         workspaceMembersOnlySelect: {
           id: "workspaceSettings.workspaceMembersOnlySelect",
@@ -317,35 +343,35 @@
             "Only workspaces with the Enterprise plan can change these settings",
         },
       });
-      function _(e, t) {
+      function N(e, t) {
         var n = Object.keys(e);
         if (Object.getOwnPropertySymbols) {
-          var r = Object.getOwnPropertySymbols(e);
+          var i = Object.getOwnPropertySymbols(e);
           t &&
-            (r = r.filter(function (t) {
+            (i = i.filter(function (t) {
               return Object.getOwnPropertyDescriptor(e, t).enumerable;
             })),
-            n.push.apply(n, r);
+            n.push.apply(n, i);
         }
         return n;
       }
-      var N = !0;
-      function E(e) {
-        return (0, v.jsx)(
-          M,
+      var E = !0;
+      function D(e) {
+        return (0, w.jsx)(
+          T,
           (function (e) {
             for (var t = 1; t < arguments.length; t++) {
               var n = null != arguments[t] ? arguments[t] : {};
               t % 2
-                ? _(Object(n), !0).forEach(function (t) {
-                    (0, r.Z)(e, t, n[t]);
+                ? N(Object(n), !0).forEach(function (t) {
+                    (0, i.Z)(e, t, n[t]);
                   })
                 : Object.getOwnPropertyDescriptors
                   ? Object.defineProperties(
                       e,
                       Object.getOwnPropertyDescriptors(n)
                     )
-                  : _(Object(n)).forEach(function (t) {
+                  : N(Object(n)).forEach(function (t) {
                       Object.defineProperty(
                         e,
                         t,
@@ -357,8 +383,8 @@
           })({}, e)
         );
       }
-      E.getLayout = function (e) {
-        return (0, v.jsx)(i.ZP, {
+      D.getLayout = function (e) {
+        return (0, w.jsx)(r.ZP, {
           mobilePageTitle: "Settings",
           requireAdminPrivileges: !0,
           children: e,
@@ -369,13 +395,13 @@
       "use strict";
       n.d(t, {
         $V: function () {
-          return f;
-        },
-        A4: function () {
           return p;
         },
+        A4: function () {
+          return f;
+        },
         Li: function () {
-          return b;
+          return j;
         },
         W8: function () {
           return y;
@@ -384,48 +410,48 @@
           return h;
         },
         i9: function () {
-          return m;
+          return b;
         },
         qQ: function () {
           return x;
         },
         vi: function () {
-          return j;
+          return m;
         },
       });
-      var r,
-        i,
+      var i,
+        r,
         s,
         a,
         o,
         l,
         c,
-        u,
-        d = n(90038),
+        d,
+        u = n(90038),
         g = n(21389),
-        f = g.Z.div(
-          r ||
-            (r = (0, d.Z)([
+        p = g.Z.div(
+          i ||
+            (i = (0, u.Z)([
               "text-base text-token-text-primary flex flex-col gap-10 my-10",
             ]))
         ),
-        p = g.Z.div(i || (i = (0, d.Z)(["text-xl font-medium"]))),
-        h = g.Z.div(s || (s = (0, d.Z)(["text-base font-medium"]))),
+        f = g.Z.div(r || (r = (0, u.Z)(["text-xl font-medium"]))),
+        h = g.Z.div(s || (s = (0, u.Z)(["text-base font-medium"]))),
         x = g.Z.div(
           a ||
-            (a = (0, d.Z)([
+            (a = (0, u.Z)([
               "flex flex-col divide-y divide-token-border-light border-y border-token-border-light",
             ]))
         ),
-        m = g.Z.div(
+        b = g.Z.div(
           o ||
-            (o = (0, d.Z)([
+            (o = (0, u.Z)([
               "max-w-xl flex flex-col gap-3 w-full m-auto md:m-0",
             ]))
         ),
-        b = g.Z.div(l || (l = (0, d.Z)(["text-sm"]))),
-        j = g.Z.div(c || (c = (0, d.Z)(["flex justify-between py-4 gap-4"]))),
-        y = g.Z.hr(u || (u = (0, d.Z)(["max-w-xl border-token-border-light"])));
+        j = g.Z.div(l || (l = (0, u.Z)(["text-sm"]))),
+        m = g.Z.div(c || (c = (0, u.Z)(["flex justify-between py-4 gap-4"]))),
+        y = g.Z.hr(d || (d = (0, u.Z)(["max-w-xl border-token-border-light"])));
     },
     16118: function (e, t, n) {
       (window.__NEXT_P = window.__NEXT_P || []).push([
@@ -443,4 +469,4 @@
       (_N_E = e.O());
   },
 ]);
-//# sourceMappingURL=settings-839d7c4366477c11.js.map
+//# sourceMappingURL=settings-65cfc8af93b61cd0.js.map